# doc/api_doc.json
{
  "consumes": [
    "application/json",
    "application/xml"
  ],
  "produces": [
    "application/json",
    "application/xml"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "REST API for Thola.\n\nFor more information look at our Github : https://github.com/inexio/thola",
    "title": "Thola",
    "license": {
      "name": "BSD",
      "url": "https://github.com/inexio/thola/blob/main/LICENSE"
    },
    "version": "1.0.0"
  },
  "host": "localhost:8237",
  "paths": {
    "/check/identify": {
      "post": {
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "tags": [
          "check"
        ],
        "summary": "Checks if identify matches the expectations.",
        "operationId": "checkIdentify",
        "parameters": [
          {
            "description": "Request to process.",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CheckIdentifyRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the response.",
            "schema": {
              "$ref": "#/definitions/CheckResponse"
            }
          },
          "400": {
            "description": "Returns an error with more details in the body.",
            "schema": {
              "$ref": "#/definitions/OutputError"
            }
          }
        }
      }
    },
...
    "/check/snmp": {
      "post": {
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "tags": [
          "check"
        ],
        "summary": "Checks SNMP availability.",
        "operationId": "checkSNMP",
        "parameters": [
          {
            "description": "Request to process.",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CheckSNMPRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the response.",
            "schema": {
              "$ref": "#/definitions/CheckResponse"
            }
          },
          "400": {
            "description": "Returns an error with more details in the body.",
            "schema": {
              "$ref": "#/definitions/OutputError"
            }
          }
        }
      }
    },
    "/check/thola-server": {
      "post": {
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "tags": [
          "check"
        ],
        "summary": "Check existence of thola servers.",
        "operationId": "checkTholaServer",
        "parameters": [
          {
            "description": "Request to process.",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CheckTholaServerRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the response.",
            "schema": {
              "$ref": "#/definitions/CheckResponse"
            }
          },
          "400": {
            "description": "Returns an error with more details in the body.",
            "schema": {
              "$ref": "#/definitions/OutputError"
            }
          }
        }
      }
    },
    "/check/ups": {
      "post": {
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "tags": [
          "check"
        ],
        "summary": "Checks whether a UPS device has its main voltage applied.",
        "operationId": "checkUPS",
        "parameters": [
          {
            "description": "Request to process.",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CheckUPSRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the response.",
            "schema": {
              "$ref": "#/definitions/CheckResponse"
            }
          },
          "400": {
            "description": "Returns an error with more details in the body.",
            "schema": {
              "$ref": "#/definitions/OutputError"
            }
          }
        }
      }
    },
    "/identify": {
      "post": {
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/json\u003c",
          "application/xml"
        ],
        "tags": [
          "identify"
        ],
        "summary": "Identifies a device.",
        "operationId": "identify",
        "parameters": [
          {
            "description": "Request to process.",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/IdentifyRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the device which was found.",
            "schema": {
              "$ref": "#/definitions/IdentifyResponse"
            }
          },
          "400": {
            "description": "Returns a string that the request was formatted wrong."
          },
          "404": {
            "description": "Returns a string that no device was found."
          }
        }
      }
    },
    "/read/available-components": {
      "post": {
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "tags": [
          "read"
        ],
        "summary": "Returns the available components for the device.",
        "operationId": "readAvailableComponents",
        "parameters": [
          {
            "description": "Request to process.",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ReadAvailableComponentsRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the response.",
            "schema": {
              "$ref": "#/definitions/ReadAvailableComponentsResponse"
            }
          },
          "400": {
            "description": "Returns an error with more details in the body.",
            "schema": {
              "$ref": "#/definitions/OutputError"
            }
          }
        }
      }
    },
    "/read/count-interfaces": {
      "post": {
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "tags": [
          "read"
        ],
        "summary": "Counts the interfaces of a device.",
        "operationId": "readCountInterfaces",
        "parameters": [
          {
            "description": "Request to process.",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ReadCountInterfacesRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the response.",
            "schema": {
              "$ref": "#/definitions/ReadCountInterfacesResponse"
            }
          },
          "400": {
            "description": "Returns an error with more details in the body.",
            "schema": {
              "$ref": "#/definitions/OutputError"
            }
          }
        }
      }
    },
    "/read/cpu-load": {
      "post": {
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "tags": [
          "read"
        ],
        "summary": "Read out the CPU load of a device.",
        "operationId": "readCPULoad",
        "parameters": [
          {
            "description": "Request to process.",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ReadCPULoadRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the response.",
            "schema": {
              "$ref": "#/definitions/ReadCPULoadResponse"
            }
          },
          "400": {
            "description": "Returns an error with more details in the body.",
            "schema": {
              "$ref": "#/definitions/OutputError"
            }
          }
        }
      }
    },
    "/read/interfaces": {
      "post": {
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "tags": [
          "read"
        ],
        "summary": "Reads out data of the interfaces of a device.",
        "operationId": "readInterfaces",
        "parameters": [
          {
            "description": "Request to process.",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ReadInterfacesRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the response.",
            "schema": {
              "$ref": "#/definitions/ReadInterfacesResponse"
            }
          },
          "400": {
            "description": "Returns an error with more details in the body.",
            "schema": {
              "$ref": "#/definitions/OutputError"
            }
          }
        }
      }
    },
    "/read/memory-usage": {
      "post": {
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "tags": [
          "read"
        ],
        "summary": "Read out the memory usage of a device.",
        "operationId": "readMemoryUsage",
        "parameters": [
          {
            "description": "Request to process.",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ReadMemoryUsageRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the response.",
            "schema": {
              "$ref": "#/definitions/ReadMemoryUsageResponse"
            }
          },
          "400": {
            "description": "Returns an error with more details in the body.",
            "schema": {
              "$ref": "#/definitions/OutputError"
            }
          }
        }
      }
    },
    "/read/ups": {
      "post": {
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "tags": [
          "read"
        ],
        "summary": "Reads out UPS data of a device.",
        "operationId": "readUPS",
        "parameters": [
          {
            "description": "Request to process.",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ReadUPSRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the response.",
            "schema": {
              "$ref": "#/definitions/ReadUPSResponse"
            }
          },
          "400": {
            "description": "Returns an error with more details in the body.",
            "schema": {
              "$ref": "#/definitions/OutputError"
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "CheckIdentifyRequest": {
      "description": "CheckIdentifyRequest is a the request struct for the check identify request.",
      "type": "object",
      "title": "CheckIdentifyRequest",
      "properties": {
        "device_data": {
          "$ref": "#/definitions/DeviceData"
        },
        "expectations": {
          "$ref": "#/definitions/Device"
        },
        "json_metrics": {
          "type": "boolean",
          "x-go-name": "JSONMetrics"
        },
        "model_diff_warning": {
          "type": "boolean",
          "x-go-name": "ModelDiffWarning"
        },
        "model_series_diff_warning": {
          "type": "boolean",
          "x-go-name": "ModelSeriesDiffWarning"
        },
        "os_diff_warning": {
          "type": "boolean",
          "x-go-name": "OsDiffWarning"
        },
        "os_version_diff_warning": {
          "type": "boolean",
          "x-go-name": "OsVersionDiffWarning"
        },
        "print_performance_data": {
          "type": "boolean",
          "x-go-name": "PrintPerformanceData"
        },
        "serial_number_diff_warning": {
          "type": "boolean",
          "x-go-name": "SerialNumberDiffWarning"
        },
        "timeout": {
          "description": "Timeout for the request (0 =\u003e no timeout)",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Timeout"
        },
        "vendor_diff_warning": {
          "type": "boolean",
          "x-go-name": "VendorDiffWarning"
        }
      },
      "x-go-package": "github.com/inexio/thola/core/request"
    },
...
    "CheckResponse": {
      "description": "CheckResponse is a generic response struct for the check plugin format.",
      "type": "object",
      "title": "CheckResponse",
      "properties": {
        "messages": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OutputMessage"
          },
          "x-go-name": "Messages"
        },
        "performance_data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/PerformanceDataPointInfo"
          },
          "x-go-name": "PerformanceData"
        },
        "raw_output": {
          "type": "string",
          "x-go-name": "RawOutput"
        },
        "status_code": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "StatusCode"
        }
      },
      "x-go-package": "github.com/inexio/thola/core/request"
    },
    "CheckSNMPRequest": {
      "description": "CheckSNMPRequest is a the request struct for the check snmp request.",
      "type": "object",
      "title": "CheckSNMPRequest",
      "properties": {
        "device_data": {
          "$ref": "#/definitions/DeviceData"
        },
        "json_metrics": {
          "type": "boolean",
          "x-go-name": "JSONMetrics"
        },
        "print_performance_data": {
          "type": "boolean",
          "x-go-name": "PrintPerformanceData"
        },
        "timeout": {
          "description": "Timeout for the request (0 =\u003e no timeout)",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Timeout"
        }
      },
      "x-go-package": "github.com/inexio/thola/core/request"
    },
    "CheckTholaServerRequest": {
      "description": "CheckTholaServerRequest is the request struct for the check thola server request.",
      "type": "object",
      "title": "CheckTholaServerRequest",
      "properties": {
        "json_metrics": {
          "type": "boolean",
          "x-go-name": "JSONMetrics"
        },
        "print_performance_data": {
          "type": "boolean",
          "x-go-name": "PrintPerformanceData"
        },
        "timeout": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Timeout"
        }
      },
      "x-go-package": "github.com/inexio/thola/core/request"
    },
...
    "CheckUPSRequest": {
      "description": "CheckUPSRequest is a the request struct for the check ups request.",
      "type": "object",
      "title": "CheckUPSRequest",
      "properties": {
        "batteryCurrentThresholds": {
          "$ref": "#/definitions/CheckThresholds"
        },
        "batteryTemperatureThresholds": {
          "$ref": "#/definitions/CheckThresholds"
        },
        "currentLoadThresholds": {
          "$ref": "#/definitions/CheckThresholds"
        },
        "device_data": {
          "$ref": "#/definitions/DeviceData"
        },
        "json_metrics": {
          "type": "boolean",
          "x-go-name": "JSONMetrics"
        },
        "print_performance_data": {
          "type": "boolean",
          "x-go-name": "PrintPerformanceData"
        },
        "rectifierCurrentThresholds": {
          "$ref": "#/definitions/CheckThresholds"
        },
        "systemVoltageThresholds": {
          "$ref": "#/definitions/CheckThresholds"
        },
        "timeout": {
          "description": "Timeout for the request (0 =\u003e no timeout)",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Timeout"
        }
      },
      "x-go-package": "github.com/inexio/thola/core/request"
    },
...
    "IdentifyRequest": {
      "description": "IdentifyRequest is a the request struct for the identify request.",
      "type": "object",
      "title": "IdentifyRequest",
      "properties": {
        "device_data": {
          "$ref": "#/definitions/DeviceData"
        },
        "timeout": {
          "description": "Timeout for the request (0 =\u003e no timeout)",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Timeout"
        }
      },
      "x-go-package": "github.com/inexio/thola/core/request"
    },
...
    "OutputError": {
      "description": "OutputError embeds all error messages which occur in requests on the API.",
      "type": "object",
      "title": "OutputError",
      "properties": {
        "error": {
          "type": "string",
          "x-go-name": "Error"
        }
      },
      "x-go-package": "github.com/inexio/thola/core/tholaerr"
    },
...
  },
  "securityDefinitions": {
    "basicAuth": {
      "type": "basic"
    }
  },
  "security": [
    {
      "basicAuth": []
    }
  ]
}